Convert ButtonElement to front/back
